Fix multiple GameState instances
Somehow the GameState class was being constructed more than once. I'm
not exactly sure how this happens because the MutationObserver should be
disconnected immediately once the object is created?

Add a guard in both the GameState constructor and handleGameStore to
prevent this from happening.

Multiple game states result in intermediate selections being transmitted
while updating a cell because the other GameState's mutex is not locked.